+++ app/Http/Controllers/HodTimesheetController.php
@@ -30,7 +30,7 @@ class HodTimesheetController extends Controller
     {
         $this->authorizeDepartment($timesheet);
         abort_unless($timesheet->status === 'submitted', 422);
-        abort_if($timesheet->user_id === auth()->id() && ! auth()->user()->isAdminLike(), 403, 'HOD cannot approve their own timesheet.');
+        abort_if((int) $timesheet->user_id === (int) auth()->id() && ! auth()->user()->isAdminLike(), 403, 'HOD cannot approve their own timesheet.');
 
         $old = $timesheet->toArray();
         $timesheet->update([
@@ -50,7 +50,7 @@ class HodTimesheetController extends Controller
     {
         $this->authorizeDepartment($timesheet);
         abort_unless($timesheet->status === 'submitted', 422);
-        abort_if($timesheet->user_id === auth()->id() && ! auth()->user()->isAdminLike(), 403, 'HOD cannot reject their own timesheet.');
+        abort_if((int) $timesheet->user_id === (int) auth()->id() && ! auth()->user()->isAdminLike(), 403, 'HOD cannot reject their own timesheet.');
 
         $old = $timesheet->toArray();
         $timesheet->update([
@@ -94,6 +94,6 @@ class HodTimesheetController extends Controller
             return;
         }
 
-        abort_unless($timesheet->department_id === auth()->user()->department_id, 403);
+        abort_unless((int) $timesheet->department_id === (int) auth()->user()->department_id, 403);
     }
 }
